feat: adjustable scroll targets for docs (#34)
Scroll targets using the :target pseudoelement, with a negative margin and height to make it possible for us to adjust the offset from the top of the viewport when a user clicks on TOC links to scroll to a heading within a docs page. See the following before/after comparing the currently deployed site at openfga.dev vs. the branch used for this PR:

![scrolling](https://user-images.githubusercontent.com/6372810/173067316-62aa82f4-7202-4f87-9eec-fbaffe9c4c84.gif)